requirements.gpu-cu75.txt
pyyaml
mxnet-cu75==0.12.1
numpy>=1.12
typing
